print all elements in a linked list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t count;
+	unsigned int count;
 
 	count = 0;
